feat: add reusable ProtectedRoute
